Refactor : SeacrchBar debouncing 적용 및 컴포넌트 분리
